+++ src/Ema/Helper/FileSystem.hs
@@ -1,7 +1,0 @@
--- | TODO: Deprecate this module!
-module Ema.Helper.FileSystem
-  ( module X,
-  )
-where
-
-import System.UnionMount as X